refactor: 검색바 분리 및 속성 추가
[#175]
