--- frontend/src/components/ui/StationSearch/index.ts
@@ -1,3 +1,0 @@
-import StationSearch from './StationSearch';
-
-export default StationSearch;
