+++ frontend/src/components/ui/StationSearchBar/index.ts
@@ -1,0 +1,3 @@
+import StationSearchBar from './StationSearchBar';
+
+export default StationSearchBar;
